Added Ability to solve Riddles
Things added:

Added the ability to find a riddle
Attempt to solve the riddle
Have the player data save to an OS independent file directory (this will change when we do our backend)

Issues:
The only problem is that you can click anywhere on the screen when the item shows up and it will take you to the riddle solver. Even the buttons.

Side Notes:
Currently, the player can guess the riddle up to three tries before losing points.
The displays will show up accordingly in the form of red messages to the user